+++ afero-sdk-softhub/src/main/java/io/afero/sdk/softhub/AferoSofthub.java
@@ -72,6 +72,23 @@ public class AferoSofthub {
     private PublishSubject<AferoSofthub> mStartSubject;
     private final PublishSubject<NotificationCallback.CompleteReason> mCompleteSubject = PublishSubject.create();
     private final PublishSubject<String> mAssociateSubject = PublishSubject.create();
+    private final PublishSubject<SetupModeDeviceInfo> mSetupModeDeviceSubject = PublishSubject.create();
+
+    public class SetupModeDeviceInfo {
+
+        public final String deviceId;
+        public final String associationId;
+
+        private SetupModeDeviceInfo() {
+            deviceId = null;
+            associationId = null;
+        }
+
+        private SetupModeDeviceInfo(String deviceId, String associationId) {
+            this.deviceId = deviceId;
+            this.associationId = associationId;
+        }
+    }
 
     private final Action0 mStartOnSubscribe = new Action0() {
         @Override
@@ -182,6 +199,10 @@ public class AferoSofthub {
 
     public Observable<NotificationCallback.CompleteReason> observeCompletion() {
         return mCompleteSubject;
+    }
+
+    public Observable<SetupModeDeviceInfo> observeSetupModeDevices() {
+        return mSetupModeDeviceSubject;
     }
 
     Observable<String> observeAssociation() {
@@ -332,14 +353,22 @@ public class AferoSofthub {
                 public void onError(Throwable e) {
                     AfLog.e("AferoSofthub startup error - deviceAssociate failed");
                     AfLog.e(e);
+
+                    Hubby.secureHubAssociationCompleted(Hubby.AssociationStatus.FAILED_PERMANENT);
+
                     startError(e);
                 }
 
                 @Override
                 public void onNext(DeviceAssociateResponse response) {
+                    Hubby.secureHubAssociationCompleted(Hubby.AssociationStatus.SUCCESS);
                     mAssociateSubject.onNext(response.deviceId);
                 }
             });
+    }
+
+    private void onSetupModeDeviceDetected(String deviceId, String assId, long profileVersion) {
+        mSetupModeDeviceSubject.onNext(new SetupModeDeviceInfo(deviceId, assId));
     }
 
     private void onOtaStateChange(String deviceId, OtaCallback.OtaState otaState, int offset, int total) {
@@ -497,6 +526,15 @@ public class AferoSofthub {
             AferoSofthub hub = mRef.get();
             if (hub != null) {
                 hub.onSecureHubAssociationNeeded(assId);
+            }
+        }
+
+        @Override
+        public void setupModeDeviceDetected(String deviceId, String assId, long profileVersion) {
+            AfLog.d("HubbyNotificationCallback.secureHubAssociationNeeded: assId=" + assId);
+            AferoSofthub hub = mRef.get();
+            if (hub != null) {
+                hub.onSetupModeDeviceDetected(deviceId, assId, profileVersion);
             }
         }